Backend/side_link/dev-requirements.txt
asgiref==3.8.1
autopep8==2.3.2
bleach==6.2.0
coverage==7.8.2
Django==5.1.7
django-cors-headers==4.7.0
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gprof2dot==2025.4.14
numpy==2.2.5
pandas==2.2.3
pillow==11.1.0
pycodestyle==2.13.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
webencodings==0.5.1

